chore: adjust StopLoadingButton style
