+++ apps/mobile/src/components/Button/type.ts
@@ -1,0 +1,7 @@
+import type { PresetColorKey } from '@/theme/interface';
+
+export type ButtonType = 'primary' | 'default' | 'text' | 'link' | 'dashed';
+export type ButtonVariant = 'outlined' | 'dashed' | 'solid' | 'filled' | 'text' | 'link';
+export type ButtonColor = 'default' | 'primary' | 'danger' | PresetColorKey;
+export type ButtonSize = 'small' | 'middle' | 'large';
+export type ButtonShape = 'default' | 'circle';
