+++ apps/mobile/src/components/index.ts
@@ -2,6 +2,7 @@ export * from '../theme/ThemeProvider';
 export { default as ActionIcon } from './ActionIcon';
 export { default as Avatar } from './Avatar';
 export { default as Button } from './Button';
+export * from './Button';
 export * from './CapsuleTabs';
 export { default as ColorScales } from './ColorScales';
 export { default as ColorSwatches } from './ColorSwatches';
