--- apps/mobile/src/features/chat/actions/SenderBtn/StopLoadingIcon/style.ts
@@ -1,19 +1,0 @@
-import { createStyles } from '@/theme';
-
-export const useStyles = createStyles((_, size: number) => ({
-  container: {
-    height: size,
-    position: 'relative',
-    width: size,
-  },
-  rotatingContainer: {
-    height: size,
-    position: 'absolute',
-    width: size,
-  },
-  staticContainer: {
-    height: size,
-    position: 'absolute',
-    width: size,
-  },
-}));
