+++ apps/mobile/src/features/chat/actions/SenderBtn/index.tsx
@@ -1,5 +1,5 @@
 import { ArrowUp } from 'lucide-react-native';
-import StopLoadingIcon from './StopLoadingIcon';
+import StopLoadingButton from './StopLoadingButton';
 import { useChat } from '@/hooks/useChat';
 import { Button } from '@/components';
 
@@ -7,7 +7,7 @@ const SenderBtn = () => {
   const { handleSubmit, isLoading, canSend, stopGenerating } = useChat();
 
   return isLoading ? (
-    <Button icon={<StopLoadingIcon />} onPress={stopGenerating} shape="circle" />
+    <StopLoadingButton onPress={stopGenerating} />
   ) : (
     <Button
       icon={<ArrowUp />}
